fix: 초기 파일 패키지 위치대로 수정
