--- src/main/kotlin/domain/result/FinalResult.kt
@@ -1,6 +1,0 @@
-package domain.result
-
-data class FinalResult(
-    val winningData: Map<Rank, Int>,
-    val rateOfReturn: String
-)
