+++ src/main/kotlin/lotto/constants/Constants.kt
@@ -1,4 +1,4 @@
-package constants
+package lotto.constants
 
 object Constants {
     const val INPUT_PURCHASE_AMOUNT = "구입금액을 입력해 주세요."
