+++ src/main/kotlin/lotto/domain/FinalResult.kt
@@ -1,0 +1,6 @@
+package lotto.domain
+
+data class FinalResult(
+        val winningData: Map<Rank, Int>,
+        val rateOfReturn: String
+)
