+++ src/main/kotlin/lotto/domain/Lotto.kt
@@ -1,4 +1,4 @@
-package domain.lotto
+package lotto.domain
 
 class Lotto(private val numbers: List<Int>) {
