+++ src/main/kotlin/lotto/domain/Money.kt
@@ -1,3 +1,3 @@
-package domain.purchase
+package lotto.domain
 
 data class Money(val amount: Int)
