+++ src/test/kotlin/lotto/LottoTest.kt
@@ -1,34 +1,12 @@
 package lotto
 
-import domain.lotto.Lotto
+import lotto.domain.Lotto
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 
 
 class LottoTest {
-
-    lateinit var lotto: Lotto
-    lateinit var inputLotties: MutableList<List<Int>>
-    lateinit var expectedResult: MutableList<List<Int>>
-
-    @BeforeEach
-    fun setup() {
-        inputLotties = mutableListOf(
-            listOf(21, 8, 23, 43, 41, 42),
-            listOf(5, 38, 16, 32, 11, 3),
-            listOf(35, 7, 44, 16, 36, 11),
-            listOf(8, 42, 31, 1, 11, 41),
-        )
-
-        expectedResult = mutableListOf(
-            listOf(8, 21, 23, 41, 42, 43),
-            listOf(3, 5, 11, 16, 32, 38),
-            listOf(7, 11, 16, 35, 36, 44),
-            listOf(1, 8, 11, 31, 41, 42),
-        )
-    }
 
     @Test
     fun `로또 번호의 개수가 6개가 넘어가면 예외가 발생한다`() {
@@ -47,8 +25,22 @@ class LottoTest {
 
     @Test
     fun `구매한 로또 정렬 테스트`() {
+        val inputLotties = mutableListOf(
+            listOf(21, 8, 23, 43, 41, 42),
+            listOf(5, 38, 16, 32, 11, 3),
+            listOf(35, 7, 44, 16, 36, 11),
+            listOf(8, 42, 31, 1, 11, 41),
+        )
+
+        val expectedResult = mutableListOf(
+            listOf(8, 21, 23, 41, 42, 43),
+            listOf(3, 5, 11, 16, 32, 38),
+            listOf(7, 11, 16, 35, 36, 44),
+            listOf(1, 8, 11, 31, 41, 42),
+        )
+
         inputLotties.zip(expectedResult) { inputItem, expectedItem ->
-            lotto = Lotto(inputItem)
+            val lotto = Lotto(inputItem)
             val actualResult = lotto.sortedRandomNumbers()
             assertThat(actualResult).isEqualTo(expectedItem)
         }
