+++ src/test/kotlin/lotto/NumberPickerTest.kt
@@ -1,6 +1,6 @@
-package domain
+package lotto
 
-import domain.lotto.NumberPicker
+import lotto.domain.NumberPicker
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
